Move retrieving wallet and chain from MetaMask in a separate file
